Rename payload shared to base
Make the name consistent.

Suggested-by: vponcova@redhat.com
