+++ pyanaconda/modules/payload/base/utils.py
