+++ pyanaconda/payload/__init__.py
@@ -39,7 +39,7 @@ from pyanaconda.payload import utils as payload_utils
 from pyanaconda.payload.install_tree_metadata import InstallTreeMetadata
 from pyanaconda.payload.requirement import PayloadRequirements
 from pyanaconda.product import productName, productVersion
-from pyanaconda.modules.payload.shared.initialization import PrepareSystemForInstallationTask, \
+from pyanaconda.modules.payload.base.initialization import PrepareSystemForInstallationTask, \
     CopyDriverDisksFilesTask
 
 from pykickstart.parser import Group
